--- components/browser/storage-sync/src/test/java/mozilla/components/browser/storage/sync/TestUtil.kt
@@ -1,9 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-package mozilla.components.browser.storage.sync
-
-import java.util.Date
-
-fun Date.asSeconds() = time / BaseGleanSyncPing.MILLIS_PER_SEC
