+++ components/support/sync-telemetry/README.md
@@ -1,18 +1,20 @@
-# [Android Components](../../../README.md) > Browser > Sync Storage
+# [Android Components](../../../README.md) > Support > Sync Telemetry
 
-A syncable implementation of `concept-storage` backed by [application-services' Places lib](https://github.com/mozilla/application-services).
+Sync telemetry helper functionality and Glean ping descriptions to be shared between sync-related components.
 
 ## Before using this component
 Products sending telemetry and using this component *must request* a data-review following [this process](https://wiki.mozilla.org/Firefox/Data_Collection).
 This component provides data collection using the [Glean SDK](https://mozilla.github.io/glean/book/index.html).
-The list of metrics being collected is available in the [metrics documentation](../../support/telemetry/docs/metrics.md).
+The list of metrics being collected is available in the [metrics documentation](docs/metrics.md).
+
+## Usage
 
 ### Setting up the dependency
 
 Use Gradle to download the library from [maven.mozilla.org](https://maven.mozilla.org/) ([Setup repository](../../../README.md#maven-repository)):
 
 ```Groovy
-implementation "org.mozilla.components:browser-storage-sync:{latest-version}"
+implementation "org.mozilla.components:support-sync-telemetry:{latest-version}"
 ```
 
 ## License
