+++ components/support/sync-telemetry/docs/metrics.md
