+++ components/support/sync-telemetry/metrics.yaml
