+++ components/support/sync-telemetry/pings.yaml
